Background correction performed using Dataframe . Processing through dictionary removed.
