+++ corna/constants.py
@@ -57,6 +57,9 @@ PARENT_NUM_LABELED_ATOMS= 'Number of labeled atoms parent'
 DAUGHTER_NUM_LABELED_ATOMS= 'Number of labeled atoms daughter'
 PARENT_MASS_ISO= 'isotopic_mass_parent'
 DAUGHTER_MASS_ISO= 'isotopic_mass_daughter'
+UNLABELED= 'Unlabeled'
+BACKGROUND_CORRECTED= 'Background Corrected'
+BACKGROUND_WITH_ZERO= 'Background Corrected with zero'
 PARENT_FORMULA_COL= 'Parent_formula'
 NA_LCMS = 'na_lcms'
 NA_MSMS = 'na_msms'
