+++ corna/inputs/multiquant_parser.py
@@ -1,21 +1,82 @@
 # TODO : many functions need to be documneted refer issue NCT-303
 
-from collections import namedtuple
 import os
 import warnings
 
-from datum import algorithms as dat_alg
 import pandas as pd
 
 from corna.inputs.column_conventions import multiquant
-from corna import constants
-from corna.helpers import check_column_headers, read_file, get_unique_values
-from corna.constants import INTENSITY_COL
+from corna import constants as const
+import corna.helpers as hlp
 
-validated_raw_tuple = namedtuple('validated_raw_mq', 'df logs')
-validated_metadata_tuple = namedtuple('validated_metadata_mq', 'df logs')
-metadata_mq_tuple = namedtuple('metadata_mq', 'df logs')
+def add_info_to_df(df):
+    """
+    This function adds column of parent isotopic mass, daughter isotopic mass, isotracer, no of labeled atoms
+    parent and daughter, total no of atoms parent and daughter to the dataframe.
+    """
+    msms_df= get_input_format(df)
+    isotracer= msms_df[multiquant.ISOTRACER].unique()
+    iso_elem= hlp.get_isotope_element(isotracer[0])
+    input_df=pd.DataFrame()
+    final_df=pd.DataFrame()
 
+    input_df= get_mass_and_number_of_atoms(msms_df, multiquant.PARENT_FORM, 'PARENT', isotracer[0],iso_elem, input_df)
+    final_df= get_mass_and_number_of_atoms(input_df, multiquant.FORMULA, 'DAUGHTER', isotracer[0],iso_elem, final_df)
+
+    return final_df, isotracer
+
+def get_num_labeled_atoms(isotope, isotopic_mass, molecular_mass):
+    """
+    This function returns no of labeled atoms of isotracer from isotopic mass and molecular mass of compound.
+    """
+    nat_iso = hlp.get_isotope_natural(isotope)
+    if nat_iso == isotope:
+        return 0
+    atom_excess_mass = hlp.get_isotope_mass(
+                isotope) - hlp.get_isotope_mass(nat_iso)
+    label= (isotopic_mass - molecular_mass) / atom_excess_mass
+    label.round(2)
+    number_label= label.astype(int)
+    return number_label
+
+def get_input_format(msms_df):
+    """
+    This function creates columns of parent isotopic mass, daughter isotopic mass, isotracer from 
+    the label column.
+    """
+    s=msms_df[multiquant.LABEL].apply(lambda x: x.split('_'))
+    msms_df[const.PARENT_MASS_ISO] = pd.to_numeric(s.apply(lambda x: x[1]), errors='coerce')
+    msms_df[const.DAUGHTER_MASS_ISO] = pd.to_numeric(s.apply(lambda x: x[2]), errors='coerce')
+    msms_df[multiquant.ISOTRACER]= s.apply(lambda x: x[0])   
+    msms_df.rename(columns={multiquant.PARENT_FORMULA: multiquant.PARENT_FORM}, inplace=True)
+    return msms_df
+
+def get_mass_and_number_of_atoms(df, formula_col, formula_info, iso_tracer,iso_elem, out_df):
+    if formula_info=='PARENT':
+        mol_mass_col= const.PARENT_MASS_MOL
+        iso_mass_col= const.PARENT_MASS_ISO
+        num_atom_col= const.PARENT_NUM_ATOMS
+        num_labeled_atoms_col= const.PARENT_NUM_LABELED_ATOMS
+    else:
+        mol_mass_col= const.DAUGHTER_MASS_MOL
+        iso_mass_col= const.DAUGHTER_MASS_ISO
+        num_atom_col= const.DAUGHTER_NUM_ATOMS
+        num_labeled_atoms_col= const.DAUGHTER_NUM_LABELED_ATOMS
+
+    for formula in df[formula_col].unique():
+        form_df= df[df[formula_col]==formula]
+        form_dict= hlp.parse_formula(formula)
+        try:
+            form_df.loc[:,num_atom_col] = form_dict[iso_elem]
+        except KeyError:
+            form_df.loc[:,num_atom_col] = 0
+        mass= hlp.get_mol_weight(formula)   
+        form_df.loc[:,mol_mass_col] = mass
+        form_df.loc[:,mol_mass_col]= form_df[mol_mass_col].round().astype(int)
+        out_df= out_df.append(form_df)
+
+    out_df[num_labeled_atoms_col]= get_num_labeled_atoms(iso_tracer, out_df[iso_mass_col], out_df[mol_mass_col])
+    return out_df
 
 def get_filtered_raw_mq_df(raw_mq, sample_metadata_mq):
     """
@@ -38,10 +99,10 @@ def get_filtered_raw_mq_df(raw_mq, sample_metadata_mq):
     """
     # :TODO: update doc with exceptions info
     try:
-        raw_filename_set = get_set_from_df_column(raw_mq, constants.ORIGINAL_FILENAME)
-        set([constants.BACKGROUND_SAMPLE]).issubset(raw_filename_set)
-        sample_filename_set = get_set_from_df_column(sample_metadata_mq, constants.ORIGINAL_FILENAME)
-        set([constants.ORIGINAL_FILENAME]).issubset(sample_filename_set)
+        raw_filename_set = get_set_from_df_column(raw_mq, const.ORIGINAL_FILENAME)
+        set([const.BACKGROUND_SAMPLE]).issubset(raw_filename_set)
+        sample_filename_set = get_set_from_df_column(sample_metadata_mq, const.ORIGINAL_FILENAME)
+        set([const.ORIGINAL_FILENAME]).issubset(sample_filename_set)
         return raw_mq
     except Exception as e:
         raise Exception(e)
@@ -86,8 +147,8 @@ def merge_samples(merged_df, sample_metadata):
         bg_corr_col_names_sample = [multiquant.BACKGROUND, multiquant.MQ_COHORT_NAME]
         bg_corr_col_names_merged = [multiquant.MQ_COHORT_NAME]
         try:
-            check_column_headers(col_headers_sample, bg_corr_col_names_sample)
-            check_column_headers(col_headers_merged, bg_corr_col_names_merged)
+            hlp.check_column_headers(col_headers_sample, bg_corr_col_names_sample)
+            hlp.check_column_headers(col_headers_merged, bg_corr_col_names_merged)
             assert set(sample_metadata[multiquant.BACKGROUND]).issubset(set(sample_metadata[multiquant.MQ_SAMPLE_NAME]))
             merged_df = merged_df.merge(sample_metadata, how='inner',
                                         on=[multiquant.MQ_SAMPLE_NAME, multiquant.MQ_COHORT_NAME])
@@ -138,11 +199,11 @@ def get_replicates(sample_metadata, sample_name, cohort_name, background_sample)
         background_sample].map(sample_index_df[cohort_name])
     replicate_groups = []
     # std should not be present in sample_metadata
-    cohort_list = get_unique_values(sample_index_df, 'Background Cohort')
+    cohort_list = hlp.get_unique_values(sample_index_df, 'Background Cohort')
     for cohorts in cohort_list:
         newdf = sample_index_df[sample_index_df[
                                     'Background Cohort'] == cohorts]
-        replicate_groups.append(get_unique_values(newdf, background_sample))
+        replicate_groups.append(hlp.get_unique_values(newdf, background_sample))
     return replicate_groups
 
 
@@ -161,7 +222,7 @@ def merge_mq_metadata(mq_df, metdata, sample_metdata):
         col_headers = merged_data.columns.values
         bg_corr_col_names = [multiquant.BACKGROUND, multiquant.COHORT]
         try:
-            check_column_headers(col_headers, bg_corr_col_names)
+            hlp.check_column_headers(col_headers, bg_corr_col_names)
         except AssertionError:
             return merged_data, list_of_replicates, sample_background
         # consider only those sample names which are present in raw file
